Wrap only method that throws with assertThrows
Improve assertThrows changes by limiting the wrapping of methods to only
the method that actually throws the exception.

Found and fixed a few EasyMock replay problems in
ManagerReplicationCoordinatorTest that only worked with a specific test
execution order, and wouldn't work when tests were run individually.

Moved some NewTableConfigurationIT tests to NewTableConfigurationTest,
because they didn't need to be integration tests, as they are expected
to fail during construction of the NewTableConfiguration, and don't
require the NTC to be used to actually create a table on a test
instance.

Replace complicated try-catch blocks that check for e.getCause() with an
assertThrows that preserves the thrown exception and checks .getCause()
directly with a subsequent assertEquals statement.

Add a few missing try-with-resources blocks for AutoCloseable
AccumuloClients in KerberosIT
